perf: add 4-bit NF4 quantization to model_loader; remove redundant model.half()
model.half() was called every query on a model already in fp16 — removed.

Added BitsAndBytesConfig support to model_loader.load():
- cfg['quantization'] = '4bit' enables NF4 quantization (6.4GB -> 1.6GB)
- cfg['quantization'] = '8bit' enables INT8 quantization (6.4GB -> 3.2GB)
- omit for default fp16 behaviour

Quantized models skip merge_and_unload() and run PEFT inference directly.
Enabled '4bit' for UC4. Expected ~2-3x decode speedup (27 -> 60-80 tok/s).

+++ pipeline/monitoring_dashboard.py
@@ -271,7 +271,7 @@ def _answer_kvforge(query: str, cfg: dict, params: QueryRequest) -> dict:
         phase = ver.get_phase()
         lora_ckpt = ver.load().get("checkpoint_path")
         model, tokenizer = _model_loader.load(lora_ckpt)
-        model = model.half()
+        # model is already in the correct dtype (fp16 or quantized) — no .half() needed
 
         # ── Phase 3: answer directly from fine-tuned weights, no retrieval ─
         if phase >= 3:
